requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.10.6
pydantic-settings>=2.1.0,<3.0.0
pydantic_core==2.27.2

# Database
sqlalchemy>=1.4.0,<2.0.0
alembic==1.13.1
psycopg2-binary==2.9.9
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.4.7
duckdb==0.9.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.28.1
httpcore==1.0.7
requests==2.31.0
h11==0.14.0

# Data Processing
pandas==2.1.4
numpy>=1.25.2,<2.0.0
PyYAML==6.0.1
pyarrow==20.0.0

# AI/ML Libraries
anthropic>=0.3.0
# openai==1.68.2  # Claude로 대체
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.20
scikit-learn==1.3.0
scipy==1.11.1
torch>=2.0.1,<2.3.0
transformers>=4.35.0,<5.0.0

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment Configuration
python-dotenv==1.0.1

# Web Framework (Existing)
Flask==3.1.0
flask-cors==5.0.1
Werkzeug==3.1.3

# Scheduling
APScheduler==3.11.0

# Data Analysis and Visualization
streamlit==1.45.1
altair==5.5.0
matplotlib==3.7.2
matplotlib-inline==0.1.6

# Web Scraping and Parsing
beautifulsoup4==4.12.2
lxml==5.4.0
feedparser==6.0.11
pytrends==4.9.2

# Jupyter Related
ipykernel==6.28.0
ipython==8.19.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
notebook==7.0.6
notebook_shim==0.2.3
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2

# System and Utilities
psutil==5.9.7
click==8.1.8
colorama==0.4.6
tqdm==4.67.1
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
tzlocal==5.3.1
watchdog==6.0.0

# Korean Language Processing
kss==4.5.1

# Excel Processing
openpyxl==3.1.2
et-xmlfile==1.1.0

# Git
GitPython==3.1.44
gitdb==4.0.12
smmap==5.0.2

# GUI (PyQt5)
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt5Designer==5.14.1

# Image Processing
Pillow==10.0.0

# Windows Only
pywin32==306
pywinpty==2.0.12

# Web Related
tornado==6.4
websocket-client==1.7.0

# Other Dependencies
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.40.0
fqdn==1.5.1
greenlet==3.2.3
idna==3.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.9.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
mistune==3.0.2
narwhals==1.40.0
nest-asyncio==1.5.8
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
platformdirs>=2.6.0,<4.0.0
prompt-toolkit==3.0.43
protobuf==6.31.0
pure-eval==0.2.2
pycparser==2.21
pydeck==0.9.1
pyee==13.0.0
Pygments==2.17.2
pyparsing==3.0.9
python-json-logger==2.0.7
pyzmq==25.1.2
referencing==0.32.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
Send2Trash==1.8.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tenacity>=8.1.0,<9.0.0
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.1.0
uv==0.7.17
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
